requirements.txt
Pillow~=10.0.1
pygame~=2.5.2
future~=0.18.3
pyglet~=2.0.9
bcrypt~=4.0.1